move consultations hook and getWeeks function into their own files
