+++ client/src/components/Calendar/get-weeks.js
@@ -1,0 +1,6 @@
+import { CalendarService } from "../../services/calendar-service";
+
+export function getWeeks(date) {
+  const dataLoader = new CalendarService(date);
+  return dataLoader.getCalendar();
+}
